+++ src/router/modules/demo/cache-demo.ts
@@ -11,7 +11,9 @@ const cacheDemo: AppRouteRecordRaw = {
     i18nKey: t('menu.CacheDemo'),
     icon: 'other',
     order: 3,
-    extra: 'new',
+    extra: {
+      label: 'new',
+    },
   },
 }