+++ src/router/modules/demo/template-hooks.ts
@@ -11,6 +11,9 @@ const axios: AppRouteRecordRaw = {
     i18nKey: t('menu.TemplateHooks'),
     icon: 'other',
     order: 1,
+    extra: {
+      label: 'new',
+    },
   },
 }